working on caching

+++ imapclient/imapclient/imap/session/base.cs
@@ -17,7 +17,7 @@ namespace work.bacome.imapclient
 
             private readonly cIMAPCallbackSynchroniser mSynchroniser;
             private readonly Action<cMailboxId, cUID, cTrace.cContext> mMessageExpunged;
-            private readonly Action<cMailboxId, uint, cTrace.cContext> mSetMailboxUIDValidity;
+            private readonly Action<cMailboxId, long, cTrace.cContext> mSetMailboxUIDValidity;
             private readonly fIMAPCapabilities mIgnoreCapabilities;
             private readonly fMailboxCacheDataItems mMailboxCacheDataItems;
             private readonly cBatchSizer mFetchCacheItemsSizer;
@@ -50,7 +50,7 @@ namespace work.bacome.imapclient
             // (note for when adding more: they need to be disposed)
 
             public cSession(
-                cIMAPCallbackSynchroniser pSynchroniser, Action<cMailboxId, cUID, cTrace.cContext> pMessageExpunged, Action<cMailboxId, uint, cTrace.cContext> pSetMailboxUIDValidity,
+                cIMAPCallbackSynchroniser pSynchroniser, Action<cMailboxId, cUID, cTrace.cContext> pMessageExpunged, Action<cMailboxId, long, cTrace.cContext> pSetMailboxUIDValidity,
                 cBatchSizerConfiguration pNetworkWriteConfiguration,
                 fIMAPCapabilities pIgnoreCapabilities, fMailboxCacheDataItems pMailboxCacheDataItems,
                 cBatchSizerConfiguration pFetchCacheItemsConfiguration, cBatchSizerConfiguration pFetchBodyConfiguration,
